Add 'public_addr' in ceph.conf by default
After upgrading the platform on AIO-DX, the ceph configuration
file (/etc/ceph/ceph.conf) is missing the "public_addr" field,
which causes the playbook to fail when performing backup & restore.

The problem is caused because this field is defined in the ceph
puppet, only in the monitor configuration, and during the upgrade
it does not enter this part, precisely because the host being
upgraded is not the active controller.

To resolve this, this field is defined together with the
"host" field in the initial ceph configuration.

  PASS: (AIO-DX) Upgrade stx.7.0 to stx.8.0
  PASS: Check that public_addr is present in ceph.conf on
        both controllers.
  PASS: Run backup & restore

Closes-Bug: 2049047

Change-Id: Ibac139fdf6d86d44f8efca1ca96f5123d639f77c
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>