Fix 'public_addr' in ceph.conf
The change made at [1] caused puppet to fail to unlock the
controller on a fresh install.

The problem was caused because public_addr was already defined
in the configuration file at the moment of configuring mon
(which is native to ceph), causing a duplicate declaration.

To fix this, 'public_addr' is set "manually" only when mon
configuration is not required.

[1]: https://review.opendev.org/c/starlingx/stx-puppet/+/905266

Test Plan:
  PASS: (AIO-DX) Upgrade stx.7.0 to stx.8.0
  PASS: (AIO-DX) Fresh install
  PASS: (AIO-DX) Backup and Restore
  PASS: (AIO-DX) Lock/Unlock/Swact

Closes-Bug: 2051287

Change-Id: I4058b663eb6aff95993edce6f2b5d561be91ef84
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>