Update SM puppet to fix rook-ceph's storage-backend-add
Update SM's puppet manifest to fix an issue in DX that was causing
an uncontrolled swact. This fix sets the correct order of execution
of the classes called by the SM in a storage-backend-add ceph-rook
command.

Test Plan:
PASS - Deploy built ISO on a DX and run storage-backend-add
PASS - Lock, Unlock and Swact
PASS - Install a rook-ceph cluster

Story: 2011055
Task: 49623

Change-Id: I9b404b1f9e3b81c461fb9a9cd364b624de4c20bd
Signed-off-by: Caio Correa <caio.correa@windriver.com>