Distribute slow REQs to 2nd USM service backend
Configure haproxy to distribute slow requests to 2nd USM service
backend.

This change adds acl option for haproxy frontend to redistribute
service requests of PUT, POST, DELETE and precheck, which are considered
as slow requests, to an alternate backend, which is serveed by a
separated USM API thread. This is to ensure general queries are
responded without delay.

This change also adds mode http to the frontends that use http method to
distribute requests, to make it obvious, even the default mode is
http.

Depends-on: https://review.opendev.org/c/starlingx/update/+/910643

TCs:
    List in depends-on commit
    passed: Provisioning systemcontroller

Story: 2010676
Task: 49649

Change-Id: I57ac49cd39082abffc94fe72834c9bb0556834ad
Signed-off-by: Bin Qian <bin.qian@windriver.com>