Fix puppet class to wipe new PV
With the change made in [1], when processing more than one
nova-local, puppet fails with "Duplicate declaration", due
to the exec "vgchange -an nova-local".

To resolve this, the variable $name was added, so that it
becomes dynamic.

Furthermore, the review mentioned above deactivate VG to perform
wiping, however, it is not activated again after that. So to
resolve this, another exec was added, so that VG is activated
after wipe new PG.

[1]: https://review.opendev.org/c/starlingx/stx-puppet/+/863871

Test Plan:
SX:  Delete instances fs and add 4 nova-local
     B&R with nova-local instead of instances

DX+: Add 4 nova-local in compute-0
     B&R with nova-local created

STD: Add 4 nova-local in compute-0
     B&R with nova-local created

Closes-Bug: 2061526

Change-Id: I7449c5cd7199541551dccee17e22a8bda48414e1
Signed-off-by: Hediberto C Silva <hediberto.cavalcantedasilva@windriver.com>
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>