Enable optional host-fs ceph filesystem
This adds an optional ceph-lv filesystem to the cgts-vg volume group
which will be used to establish ceph specific storage on controllers and
workers. The presence of this filesystem on a particular host will be
used to determine where containerized Ceph services will be provisioned
via Rook.

This follows the established patterns set by the instances and
conversion filesystems.

Test Plan:
 - PASS: Install AIO-SX with bare-metal Ceph
 - PASS: Install AIO-DX with bare-metal Ceph
 - PASS: Install 2+2 with bare-metal Ceph
 - PASS: add/remove optional ceph host-fs on controller-0/1
 - PASS: add/remove optional ceph host-fs on worker

Change-Id: Ie14156b7abec4cc970792b7714746880499c9fe2
Story: 2011117
Task: 50037
Signed-off-by: Robert Church <robert.church@windriver.com>