Memcached using hostname
The Memcached is started by init.d before the SM starts the DNSMasq.
This results in the service not being able to resolve the FQDN and
the service starts and stop several times during the startup, logging
some error messages until the DNSMasq is started.
The Memcached service switched to use the FQDN because after a
management network reconfiguration it was using the wrong IP.
Since a management network reconfiguration is not a frequent action
the decision is to use the hostname for the normal operation and
restart the memcached after a management network reconfiguration to
fix the IP address.

Test done:
AIO-SX fresh install
AIO-DX fresh install
AIO-DX host-swact

Story: 2010722
Task: 50222

Depends-On: https://review.opendev.org/c/starlingx/config/+/920694

Change-Id: I4b0e88f3be7e06ed7e5b548a7e018ba24fa50056
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>