Use kubectl create/replace rather than apply for gnp rules
A scalability issue was discovered in a large distributed cloud
system with thousands of subclouds.

When a subcloud is added, Calico globalnetworkpolicy rules are
created to allow the subcloud and system controller to be
permitted to communicate via the firewall.

The gnp rules are generated via a file constructed by puppet
based on the available (known) routes.

What can happen with thousands of subclouds (routes + rules) is that
the file becomes so large that kubernetes cannot handle it via
a 'kubectl apply <file>' command.

The kubectl apply is trying to reconcile what the user wants
(the file) with what is present on the system.

The error message looks similar to:

The GlobalNetworkPolicy "controller-mgmt-if-gnp" is invalid:
metadata.annotations: Too long: must have at most 262144 bytes

As is expected, it is not easy to gain access to a system with
thousands of subclouds - but I have reproduced the issue by
constructing a gnp file simulating the real system.

A solution to this is to directly call kubectl create/replace
rather than apply.

Test Plan:

- Manually kubectl create/replace a file with 10000 rules.
- Using the system host-route-add command, create ~8000 routes
  (each one leads to puppet updating the gnp rules file and
   replacing the Calico rules)

Closes-Bug: #2068499

Change-Id: I99f075488c164816c1e8a6f31421a1cbe7dcc78e
Signed-off-by: Steven Webster <steven.webster@windriver.com>