Fix variable reassign on containerd.pp
During the upstream containerd v1.5 code deprecated change
a portion of containerd.pp attempts to reassign a variable
($registry). Changing that and using different variables
for both local and central registries. This variables are
later used in configuration templates.

Registry Configuration for hosts:
https://github.com/containerd/containerd/blob/main/docs/hosts.md

Test Plan:
PASS: Verify containerd config file format are as
      expected.
PASS: Installing on a subcloud in a DC environment

Closes-bug: 2070977

Change-Id: I273360f1b7607adc0ed0ef3c5815ae055f1973d2
Signed-off-by: rakshith mr <rakshith.mr@windriver.com>