Provision pxeboot service in USM upgrade
When in an USM major release deployment, the pxeboot service
is not being configured because /etc/platform flags are retained
when the host deploys the to-release ostree commit-id. As a
consequence of the service not being provisioned correctly makes
dnsmasq service to fail later when swacting.

This commit includes a condition to provision the service when an
USM upgrade is in progress.

Test Plan
PASS: (AIO-DX) stx-8 -> stx-10 major release deployment

Story: 2010676
Task: 50679

Change-Id: I0b0d260390cabd1fc38ad974550a7be7f0a00c9c
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>