Disable dcagent user creation on SystemController
It's not necessary to have the dcagent keystone user present on
system controller. Its presence can cause problems during upgrade
as it would be necessary for the dcagent user in the subcloud to
have the same password as system controller upon creation (before
dcdbsync synching upon managing the subcloud).

Test plan:
  - PASS: Build a developer ISO. Deploy a DC system with 2 system
          controllers and 1 subcloud. Verify the dcagent user is
          created only on the subcloud.
  - PASS: Manage the subcloud and verify the dcagent service is
          working as expected.
  - PASS: Perform a network reconfiguration on the subcloud. Verify
          the operation finishes successfully.

Story: 2011106
Task: 50747

Change-Id: I13deb83f2e8b602fd8d733b999cf579d4db05034
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>