Update strongswan.pp to support SX to DX migration
This commit changed the way of how the ipsec service restart command is
determined. Instead of by checking initial_config_complete file, it is
now checking the presence of its pmon configuration file. This is
because the initial_config_complete file flag already exists on the SX
system to be migrated so it will cause a wrong ipsec service restart
command, when ipsec is enabled.

This commit also moved the pmon configuration generation to the end of
configuration chain, to avoid the race condition between pmon and
systemd to start ipsec service for the first time the puppet class is
applied.

The config class in strongswan.pp will be invoked by the following
command after the controller is unlocked in the migration procedure.
This will be an new step required by IPsec.

  ipsec-client pxecontroller

Note: there will be follow-up change to migration playbook to support
      subcloud migration.

Test Plan:
PASS: DX system deployment.
PASS: Follow manual procedure to migrate a SX system to DX system,
      Verify the migration is successful.

Story: 2010940
Task: 50723

Change-Id: Ic30e5bf0605d515401e7f264c0b7109e85c4b050
Signed-off-by: Andy Ning <andy.ning@windriver.com>