Fix Ceph mds Pmon settings on compute nodes for AIO-DX+
When installing an AIO-DX+ system, puppet was configuring Pmon to
monitor ceph-mds process on every host, even the compute nodes. This
was raising alarms because the service was not configured properly and
it was failing to start.

Puppet manifest is changed to only configure Pmon to monitor Ceph mds
processes on controller nodes for AIO-DX and AIO-DX+ systems.

There are no changes for the other configurations.

Test Plan:
  PASS: Install AIO-DX system with Ceph and check the alarms
  PASS: Install AIO-DX+ system with Ceph and check the alarms
  PASS: Install AIO-SX system with Ceph and check alarms
  PASS: Install STD system with Ceph and check alarms

Partial-bug: 2076369

Change-Id: I84039c73f9d83ee8415a83942d363226b2b5d8fb
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>