Add extra frontend for USM internal endpoint
Due to changing calls from the CLI to USM 'public' to 'internal'
endpoints, there instances where deploy precheck is failing due to
circular calls between USM and sysinv.

One solution for this is to load balance in haproxy the same way as
already done when the 'public' endpoint is used, sending some of the
calls to an alternative backend.

In order to load balance in haproxy for internal endpoints, added an
extra frontend binded to the mgmt ip.

Opted to move the first backend port to (private_port + 2) to avoid
colliding w/ the admin frontend port convention for DC (private_port
+ 1), since both use the mgmt IP.

In USM case, the ports were changed to:
- frontend (internal): 5497 (private_port)
- frontend (admin):    5498 (private_port + 1)
- backend default:     5499 (private_port + 2)
- alt backend (slow):  5500 (can be arbitrary)

Test plan:
PASS: Bootstrap and unlock AIO-SX, verify haproxy.cfg to see if the
      new frontend is in place and the backend ports are correct.
PASS: Bootstrap and unlock DC + SX subcloud, verify haproxy.cfg to
      see if the new frontend is in place and the backend ports are
      correct.
PASS: Execute 'software deploy precheck', verify that the 'precheck'
      request is reaching the alternative backend.
PASS: Execute 'sw-manager sw-deploy-strategy create', verify that the
      'precheck' request is reaching the alternative backend.

Story: 2010676
Task: 50910

Depends-on: https://review.opendev.org/c/starlingx/update/+/926947

Change-Id: I1b24e0fe17be6e023ec0c2587f925fced57cbd9a
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>