Fix disabling admin network config
When switching from admin network to mgmt network, the SM admin-ipvX
service should be unmanaged and deprovisioned. A change was added to
disable the services for the dual stack support. Since the admin network
parameters in hieradata are removed when moving from admin to mgmt
network, the check to disable the admin-ipvX service was not executed.
This commit is to unmanage and deprovision the service only if the
service is found.

Test plan:
Pass: Switch from admin network to mgmt network
Pass: Switch from mgmt network to admin network

Story: 2011027
Task: 50960

Change-Id: I8941c3335e8e810414892262f87fc8862d681966
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>