Fix Ceph being unresponsive on AIO-DX standalone controller
When installing an AIO-DX setup and the standalone controller-0 is
rebooted after unlocking it and before installing/unlocking
controller-1, the Ceph gets unresponsive due to a lack of monitors
quorum. This happens because in the puppet manifest only the floating
monitor is being started and the other settings in the manifest
configuring the OSDs need Ceph to be responsive.

Starting the fixed ceph monitor on the host that has the floating
monitor also running is enough to have Ceph responsive during puppet
execution.

If the host is the standby controller, the fixed monitor will not be
started during puppet manifest application due to race conditions
when applying the Ceph configuration in runtime. The Pmon will start
the process later.

Test-Plan:
  PASS: AIO-DX - Deploy both controllers and check Ceph is reporting
        HEALTH_OK status with 3 monitors quorum.
  PASS: AIO-DX - Install and unlock only Controller-0. Check Ceph
        is responsive with 2 monitors quorum. Reboot the
        Controller-0 host and check again if Ceph is responsive with
        2 monitors quorum.
  PASS: AIO-DX - Install and unlock only Controller-0. Check Ceph
        is responsive with 2 monitors quorum. Reboot the
        Controller-0 host and check again if Ceph is responsive with
        2 monitors quorum. Continue installing Controller-1 host.
        After unlocking Controller-1, check Ceph is reporting
        HEALTH_OK status with 3 monitors quorum.
  PASS: AIO-DX - Deploy without Ceph configured. Configure Ceph in
        runtime. Check Ceph HEALTH_OK status with 3 monitors quorum.

Story: 2011122
Task: 51010

Change-Id: I105946515e14eb019ada18686f843b796b6dfdd7
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>