Prevent mtce runtime manifest from sighup'ing processes not running
The mtce runtime manifest tries to SIGHUP the mtcAgent on the standby
controller. Since the mtcAgent process is never running on the standby
this leads to a traceback and manifest execution error.

This update modifies the maintenance puppet runtime manifest to only
SIGHUP the mtcAgent and hbsAgent if these processes are found running.

Test Plan:

PASS: Verify build and install of AIO DX system
PASS: Verify a platform maintenance service parameter change(s) run
      the mtce runtime manifest which updates /etc/mtc.ini with the
      modified change.
PASS: Verify the mtcAgent on the active controller is SIGHUP'ed
PASS: Verify the mtcAgent on the standby controller is not SIGHUP'ed
PASS: Verify the mtce runtime manifest puppet log file content for
      correctness on both controllers.

Closes-Bug: 2080577
Change-Id: I0578bcffa32ccd5782ec464488b55fb37f662015
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>