Fix .ceph_started flag creation
The flag .ceph_started was being created in the ceph.pp manifest.
This flag enables SM and Pmon monitoring, but the Ceph processes might
not be initialized yet. This flag is created in ceph.sh script called
by MTC when enabling the host.

When configuring ceph backend at runtime, the MTC will not call the
ceph.sh script and the processes will not be fully started and the
flag will not be created. To solve this, the ceph.sh script will be
called at the end of puppet manifest runtime apply.

Test-Plan:
  PASS: On all setups, do a fresh install and check the flag is not
        being created during the puppet manifest apply.
        The flag must be created by the MTC.
  PASS: On all setups, do a fresh install without Ceph backend
        configured. Configure Ceph at runtime and check if the flag is
        created after Ceph is configured.

Partial-bug: https://bugs.launchpad.net/starlingx/+bug/2083056
Depends-on: https://review.opendev.org/c/starlingx/integ/+/930514

Change-Id: Ibbd7dbb41f00c2b2354eaa9a5bd8d383a3d63ac8
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>