AIO-DX: Add super-admin.conf file on standby controller for K8s v1.29
A fresh installation of Kubernetes v1.29, as well as an upgrade from
v1.28 to v1.29 in a duplex system, creates the super-admin.conf file
on the active controller but not on the standby controller.

The super-admin.conf file should exist on both controller nodes for
redundancy purposes.

This change ensures that the super-admin.conf file is generated on
the standby controller during both a fresh install of K8s v1.29 and
the upgrade from v1.28 to v1.29.

Test Plan:
PASS: Install ISO with K8s 1.29 on AIO-DX and verify that the
      super-admin.conf is present on both controllers.
PASS: Install ISO with K8s 1.28 on AIO-DX, upgrade to 1.29 and verify
      that the super-admin.conf is present on both controllers.
PASS: Install ISO with K8s 1.28 on AIO-DX, set the controller-1 as
      active, upgrade to 1.29 and verify that the super-admin.conf
      is present on both controllers.
PASS: Verify that "sudo kubeadm certs check-expiration" command
      outputs the super-admin-conf details on both controllers.

Closes-Bug: 2081769

Change-Id: I58b6a995b37b70e8b3350311ca3c89e4a008f8b7
Signed-off-by: Ramesh Kumar Sivanandam <rameshkumar.sivanandam@windriver.com>