Ownership and Permission Changes for /etc/cron*
Currently, CIS benchmark checks fail for the permission and ownership
settings of /etc/cron* and the /etc/at.allow file. The CIS
recommendation is to ensure that root is both the owner and group for
these files, and that only the owner has access to them.

This change introduces a Puppet class that ensures file permissions and
ownership for /etc/cron* and /etc/at.allow are modified according to
CIS benchmark recommendations.

Test Plan:
PASS: Build iso and deploy.
PASS: AIO-SX: Ensure that the ownership and permission of
      /etc/cron* and /etc/at.allow are set as per the CIS
      recommendation.
PASS: AIO-SX: Run the CIS benchmark and ensure that the
      controls/tests related to /etc/cron* and /etc/at.allow
      pass successfully.
PASS: AIO-SX: Verify that cron jobs are executed correctly
      after the change. Verify that there are no errors in the
      cron log at /var/log/cron.log. Specifically, check for the
      absence of any error messages or failed cron job logs.Verify
      that the memory log file rss-memory.log exists in the
      expected directory and a log of cron execution in
      /var/log/cron.log.

Story: 2011241
Task: 51170

Change-Id: Ib72caedcd21be07daa310aa6eef21d26cc8db7cb
Signed-off-by: Jagatguru Prasad Mishra <jagatguruprasad.mishra@windriver.com>