Increase HAProxy USM timeout for "slow" Requests
The software upload command currently fails due to a timeout while
awaiting the HTTP response.
This issue commonly occurs when uploading larger patches, such as
a 1GB file.
For example:
software upload test.patch

This timeout issue may occur on the client-side request, which is
addressed in the following fix:
https://review.opendev.org/c/starlingx/update/+/934084

Additionally, an adjustment is needed on the HAProxy side to increase
the timeout for "slow" USM requests (PUT,POST,DEL + precheck requests)

Test Done:

PASSED software upload on AIO-SX with a 1GB patch file with timeout 1800s
PASSED software upload on AIO-SX with a 1GB patch file with timeout 300s
       Edited the file: /etc/haproxy/haproxy.cfg
       Changed the "timeout server" setting to 300 seconds for the
        backend: alt-usm-restapi-internal
       Restarted the HAProxy service:
         sudo sm-restart service haproxy
       Apply the patch:
         software upload large.patch
       Confirmed the command failed due to HAProxy timeout,
       with the error:
           Gateway Timeout

Depends-On: https://review.opendev.org/c/starlingx/update/+/934084

Story: 2010676
Task: 51265

Change-Id: I3a520455ceb150d753262d1542e5c4e608acec99
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>