Add instrumentation log for kubeadm join command
This uses kube_command helper for logging instrumentation
of 'kubeadm join' command. This is useful in cases when the
join command fails or hits timeout. In the case of timeout,
we currently get no indication of progress or the actual failure.

The platform::kubernetes::kube_command helper function is
updated to have new optional parameter 'unless', and the
'environment' parameter is modified to pass an array instead
of a string to handle an empty array as the default.

Partial-Bug: 2084714

TEST CASES:
PASS: AIO-SX, AIO-DX, Standard, DC: Fresh install ISO.
      Verify we get file output logs in /var/log/puppet/<dir>/
      for kubeadm-join-command.log with verbose output.
PASS: AIO-DX: K8S Orchestrated upgrade

Change-Id: Id88d07a62d9bd8785227213d5e1b49fca5260084
Signed-off-by: Jim Gauld <James.Gauld@windriver.com>