Fix CLI behavior change for Helm v > 3.3.1
Helm introduced changes in the CLI behavior for the command
'helm repo add' in versions superior to 3.3.1. For reference:
https://github.com/helm/helm/issues/8771 .

This caused the code inside platform::helm::repository to return an
error when repos are updated, which in turn makes the puppet manifest
fail.

The previous behavior can be achieved by using the flag 'force-update'
introduced by Helm. However, this is not backwards compatible, so it's
usage is conditioned to the software version to decrease the chance of
issues during upgrades from stx 8.0.

Test plan:
PASS: Bootstrap DC + SX subcloud.

PASS: Switch http_port to 80 (default is 8080).
      Verified that the puppet manifest executed successfully.
      Verified that helm repos were updated.

PASS: Switch https_enabled to false.
      Verified that Horizon is accessible using the port 80.

Story: 2011266
Task: 51407

Change-Id: I5dd0a5ac1914073a6e500cd9dde602aa63b874eb
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>