Correct After directive in IPsec systemd override
The After directive in IPsec sysetmd service override
("After=network-online.target luks-fs-mgr.service") is added in the
[Service] section whereas it should be in the [Unit] section. This
causes the dependency in the After directive is ignored and not
enforced.

Also since dependencies can only be added in drop-ins,
"network-online.target" is not neccessay, as it is already in the
original [Unit] section.

Test Plan:
PASS: Deploy a multi-nodes system, check daemon.log, verify that
      strongswan-starter service started after luks-fs-mgr.service,
      and there are no "Unknown key name 'After' in section 'Service',
      ignoring" warnings for strongswan-starter.service.

Closes-Bug: 2089874

Change-Id: I29dafe526d84d2aa05c74a29d0ceb94bc0ca0b01
Signed-off-by: Andy Ning <andy.ning@windriver.com>