Ignore kubectl error output when searching firewall's GNP
This change removes the error message "Error from server (NotFound):
globalnetworkpolicies.crd.projectcalico.org [gnp name] not found" from
the operation destined to verify if the GNP exists as the result will
be used in the next line. No need to print the error message as that
can lead to wrong interpretation in the system logs.

Test Plan
[PASS] Install AIO-DX and observe that the message doesn't appear
       anymore but the firewall still is correctly installed

Story: 2011324
Task: 51531

Change-Id: I75180a7d1ff8d706d5bc6ed9b5e3d630f5e811f8
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>