Update pause image version during k8s Upgrade
As part of the Kubernetes upgrade process, the pause image version may
be updated. We need to ensure that /etc/containerd/config.toml is updated
during the kubelet upgrade step to match the version corresponding to
the new k8s release. This ensures that, during backup and restore,
the correct version of the pause image is downloaded and made available
to pods. Without this change, the upgraded pause image may not be available
during backup and restore, and the containerd config.toml would
reference an outdated version.

Test Plan:
  Pass: Upgrade k8s on AIO-SX and confirm config.toml is updated.
  PXXX: Upgrade k8s on standard configuration and confirm config.toml
        is updated.

Closes-Bug: 2095090

Change-Id: If5c35f58234db85d013dd55c443b12d7599fee03
Signed-off-by: Gleb Aronsky <gleb.aronsky@windriver.com>