Only use defined addresses for certSANs configuration
A problem was identified during the migration from AIO-SX to AIO-DX
with dual-stack (IPv4 and IPv6). The Kubernetes certSANs configuration
failed when the primary OAM address pool was configured with the
addresses of controller0 and controller1. This failure stemmed from
the absence of a corresponding secondary OAM address pool
configuration for the same controllers. Consequently, an invalid
string was generated for the config map configuration.

This change tests the defined address variables before adding them to
the certsans string.

Test Plan
=========
[PASS] install AIO-SX in dual-stack:
       - migrate config to AIO-DX
       - adjust the addresses for both OAM pools
       - unlock controller-0
       - proceed with controller-1 installation
[PASS] install AIO-SX in single-stack and migrate to AIO-DX

Closes-Bug: 2103782

Change-Id: Id022ca86d14c0bb60d691b739e383a6e92c6ab54
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>