Fix ifup error message of vlan that is already configured by kickstart
The first time apply_network_config.py runs in non controllers with mgmt
over VLAN, the pxeboot/mgmt interface is already configured by ifupdown.
When ifup is called for the label that holds the mgmt address, the
following error occurs:
Error: ipv6: address already assigned.
ifup: failed to bring up vlan10:1-22
This happens because since the /etc/network/interfaces.d/auto file does
not exist yet, all the interfaces defined by puppet are considered "new"
and are not required to be set down before ifup.
This change adds logic to check if any of the newly introduced
interfaces are already up, and if it is the case, adds them to the
down list.

Test plan

[PASS] STANDARD IPv6 full install on VirtualBox (2 storages + 1 compute)

Partial-Bug: 2103645
Change-Id: Id7968521606ea54085468b1e55798c75663e7561
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>