Save log file permission changes during upgrade
To comply with CIS benchmark requirements, the log file permissions are
set to 640 during a host upgrade. If the upgrade is rolled back, the
file permissions should be reverted to their previous state.

To facilitate this rollback, the previous permissions are stored in a
hidden file named .permission.txt. This file is used to restore the
original permissions during an upgrade rollback.

Test Plan:
PASS: Build ISO.
PASS: Deploy AIO-SX (24.09) and upgrade to next version using the
      designer iso.Verify that .permission.txt exists on controller-0
      and it contains the correct file permissions.
PASS: Deploy standard (24.09) and upgrade to next version using the
      designer iso.Verify that .permission.txt exists on all the nodes
      and it contains the correct file permissions.

Story: 2011241
Task: 51853

Change-Id: I567b0d7f557e77666da2be4a8255fdf68f8873c6
Signed-off-by: Jagatguru Prasad Mishra <jagatguruprasad.mishra@windriver.com>