Remove registry central resolution from dns puppet template
This commit removes the name resolution for 'registry.central' within
the Puppet template. The objective is to eliminate the use of the
distributed cloud role in this context. The configuration is now
replaced with the use of a service parameter DNS host record.

Test Plan
=========
For all the tests below, ensure that the address entry for
'registry.central' is not configured in the dnsmasq.conf file.

[PASS] Execute subcloud bootstrap. After deployment, ping
registry.central and verify the service-parameter-list.
registry.central must be properly configured.

[PASS] Execute subcloud bootstrap with the virtual_system flag enabled.
After deployment, ping registry.central and verify the
service-parameter-list. registry.central must be properly configured.
registry.local must also be configured.

[PASS] After subcloud deployment, modify registry.central DNS
host-record to another IP address and ping. registry.central must be
properly configured.

[PASS] Execute system controller bootstrap. Check
service-parameter-list: registry.central must not be configured.

[PASS] Execute subcloud rehoming. After rehoming, registry.central must
be updated and point to the new system controller IP address. The
address must be reachable and properly configured.

[PASS] Execute subcloud rehoming using same OAM/mgmt values. Rehoming
must be executed with no failures. registry.central must be reachable
and properly configured.

[PASS] Execute subcloud rehoming with the virtual system flag enabled.
After rehoming, registry.central must be updated and point to the new
system controller management IP address. The address must be reachable
and properly configured. registry.local must also be properly
configured.

[PASS] Execute subcloud enrollment. After enrollment, registry.central
must be updated and point to the system controller OAM IP address. The
address must be reachable and properly configured.

[PASS] Execute subcloud enrollment with the virtual system flag enabled.
After enrollment, registry.central must be updated and point to the
system controller management IP. The address must be reachable and
properly configured. registry.local must also be properly configured.

[PASS] Execute backup and restore process. registry.central should
be restored successfully.

[PASS] Execute DC fresh install. registry.central must be properly
configured for subcloud.

[PASS] Execute AIO-SX fresh install. Check service-parameter-list:
registry.central must not be configured.

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/945991

Story: 2011312
Task: 51903

Change-Id: I431cb0d780187f0ad821f32e2f3c1c21e59c79e7
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>